[GP-14] update proto (#8)
* [GP-14] update proto

* [GP-14]adjust pipeline status order

* [GP-14] test

+++ internal/model/scrpc_service.go
@@ -17,7 +17,7 @@ type RPCService struct {
 	BuildFileRelPath   string `json:"buildFileRelPath" gorm:"column:build_file_rel_path"`
 	Type               string `json:"type" gorm:"column:app_type"`
 	// CustomPort is valid only if Type == config.AppTypeHTTP
-	CustomPort uint8 `json:"customPort" gorm:"column:custom_port"`
+	CustomPort uint16 `json:"customPort" gorm:"column:custom_port"`
 	// PrefixMapping is valid only if Type == config.AppTypeHTTP
 	PrefixMapping string `json:"prefixMapping" gorm:"column:prefix_mapping"`
 }
